get_all_related_objects compat api
